refactor: Performs rebase from master
