+++ src/main/java/eu/bbmri_eric/negotiator/discovery/synchronization/DirectoryClient.java
@@ -11,7 +11,6 @@ import eu.bbmri_eric.negotiator.discovery.DiscoveryServiceRepository;
 import eu.bbmri_eric.negotiator.form.AccessForm;
 import eu.bbmri_eric.negotiator.form.repository.AccessFormRepository;
 import eu.bbmri_eric.negotiator.governance.network.Network;
-import eu.bbmri_eric.negotiator.governance.network.NetworkCreateDTO;
 import eu.bbmri_eric.negotiator.governance.network.NetworkRepository;
 import eu.bbmri_eric.negotiator.governance.organization.Organization;
 import eu.bbmri_eric.negotiator.governance.organization.OrganizationRepository;
@@ -19,6 +18,7 @@ import eu.bbmri_eric.negotiator.governance.resource.Resource;
 import eu.bbmri_eric.negotiator.governance.resource.ResourceRepository;
 import eu.bbmri_eric.negotiator.governance.resource.dto.MolgenisBiobank;
 import eu.bbmri_eric.negotiator.governance.resource.dto.MolgenisCollection;
+import eu.bbmri_eric.negotiator.governance.resource.dto.MolgenisNetwork;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.NoSuchElementException;
