+++ src/main/java/eu/bbmri_eric/negotiator/governance/resource/dto/GenericNetwork.java
@@ -1,4 +1,4 @@
-package eu.bbmri_eric.negotiator.dto.discoveryservice;
+package eu.bbmri_eric.negotiator.governance.resource.dto;
 
 public interface GenericNetwork {
