+++ src/main/java/eu/bbmri_eric/negotiator/governance/resource/dto/GenericOrganization.java
@@ -1,4 +1,4 @@
-package eu.bbmri_eric.negotiator.dto.discoveryservice;
+package eu.bbmri_eric.negotiator.governance.resource.dto;
 
 public interface GenericOrganization {
   String getId();
